+++ src/Cwaff-Misc/CwaffSynergies.cs
@@ -326,6 +326,8 @@ public static class CwaffSynergies
         NewMastery<MasteryOfForkbomb>(MASTERY_FORKBOMB, Forkbomb.ItemName);
         // Zealot gains infinite ammo and fires even while dodge rolling.
         NewMastery<MasteryOfZealot>(MASTERY_ZEALOT, Zealot.ItemName);
+        // Toothpaste projectiles create suds in a much larger radius, and toothbrush can be swung much faster.
+        NewMastery<MasteryOfToothpaste>(MASTERY_TOOTHPASTE, Toothpaste.ItemName);
       #endregion
 
         SanityCheckAllSynergiesHaveBeenInitialized();
@@ -607,6 +609,7 @@ internal class MasteryOfLeafblower           : MasteryDummyItem {}
 internal class MasteryOfWayfarer             : MasteryDummyItem {}
 internal class MasteryOfForkbomb             : MasteryDummyItem {}
 internal class MasteryOfZealot               : MasteryDummyItem {}
+internal class MasteryOfToothpaste           : MasteryDummyItem {}
 
 public enum Synergy {
     // Synergies
@@ -750,4 +753,5 @@ public enum Synergy {
     MASTERY_WAYFARER,
     MASTERY_FORKBOMB,
     MASTERY_ZEALOT,
+    MASTERY_TOOTHPASTE,
 };